requirements.txt
  GNU nano 5.8                    requirements.txt
Django==4.0.1
django-adminlte3
django-cors-headers
djangorestframework
djangorestframework-simplejwt==5.1.0
asgiref
atomicwrites
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow
pluggy==1.0.0
py==1.11.0
PyJWT
pyparsing==3.0.6
